Move App from container to components directory. Delete test, will be created later.
