+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import { StatusProvider } from './context/StatusContext';
-import App from './containers/App/App';
+import App from './components/App/App';
 // import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(
